Introduce neonvm-guest, for running payload in a container in the VM
In the new system, the "payload" is run in a container in the VM. In
the usual setup, the payload is the compute-node image built from the
main 'neon' repository, which includes "compute_ctl", pgbouncer and
other such services that are needed to run a PostgreSQL compute
endpoint. Like before, the payload is provided in a docker container
which contains a QCOW2 image. But you don't need to use vm-builder to
build the payload image anymore. The payload can use systemd to launch
the services.

The VM now always boots the same disk image called "neonvm-guest".
The neonvm-guest disk image is built into the neonvm-runner docker
image. It's based on Debian bookworm and uses systemd to launch the
services needed in all neonvm VMs: neonvmd, vector, and ssh for
debugging.  At startup, alongside launching those common services,
neonvm-guest mounts the payload image, and runs it as a container
using systemd-nspawn.

Running the payload in a container provides some isolation between the
common VM services and the real payload. It also allows using a
different OS for the payload. We're standardizing on Debian bookworm,
but as of this writing, older PostgreSQL versions still use a Debian
bullseye based image.

This is based on Em's proposal at
https://www.notion.so/neondatabase/RFC-Containers-in-VMs-dd5f4d978ec6499aba0662bd54421b91#33c2929d8880494088331400213eea90

To make all this work, there are changes to:

- neonvm-runner: attach neonvm-guest and neonvm-payload as separate
  QCOW2 disk images to the VM. It still works with legacy VM images
  too; a neonvm-guest based image is detected by looking at the QCOW2
  image name: if it's neonvm-payload.qcow rather than rootdisk.qcow,
  it uses the new system. The neonvm-guest disk image is always built
  into the neonnm-runner image, but it will go unused when launching a
  legacy VM image.

- neonvm-controller: Accept both disk.qcow2 and neonvm-payload.qcow2
  images in the VM image.

TODOs:

- Resizing swap setting disk quotas is not implemented in the new
  system yet.

- ARM support. I think the serial TTY name is wrong on ARM. Otherwise
  it should work, but is untested.

- I'd love to rewrite the way that the compute_ctl arguments are
  passed to the VM. Currently, they're put command.sh and args.sh
  files which are mounted in the VM under /neonvm/runtime/. That feels
  pretty janky, and it was hard to keep that system working with the
  new systemd-based approach. In fact I think that's broken at the
  moment.

- I originally started working on this in order to better support
  running one-time jobs in VMs, and in particular the fast_import tool
  that's part of the migration assistant feature. I'm thinking that it
  could be packaged as a different "payload" image. Or with different
  parameters to the same VM compute image. But this PR doesn't address
  that yet.

- If one of the services fails to start, should the VM be powered off?
  Which ones may fail, which ones are considered hard failures?

- I'd like to be able to use the same compute VM, with compute_ctl,
  for all PostgreSQL versions. The straightforward way to do that is
  to include the binaries of all PostgreSQL versions in the same
  payload image. Alternatively, include multiple payload images in the
  VM, and "mount" the correct one on the fly when we know which
  version is needed.

- Ideally, extensions and ad hoc compute images could be loaded
  on-demand too, while still using the VMs from the compute pool. That
  means lazily downloading the right image into the VM, after the spec
  file is received from the control plane. Need to figure out the best
  way to do that and the previous point.